feat: - Added Dot badge - Added Dot stories
