+++ src/index.ts
@@ -22,6 +22,7 @@ import Dropdown from './components/dropdown';
 import User from './components/user';
 import TextValueBadge from './components/badges/text-value-badge';
 import Badge from './components/badges/badge';
+import Dot from './components/badges/dot';
 
 // Navigation
 import Navigation from './components/navigation/container';
@@ -70,6 +71,7 @@ export {
   User,
   TextValueBadge,
   Badge,
+  Dot,
   // Navigation
   Navigation,
   NavigationItem,
