+++ src/theme/badges/dot.ts
@@ -1,0 +1,10 @@
+const variants = {
+  black: {
+    bg: 'black',
+  },
+  green: {
+    bg: 'green',
+  },
+};
+
+export default variants;
