+++ src/theme/badges/index.ts
@@ -1,9 +1,11 @@
 import keyValue from './keyValue';
 import primary from './primary';
+import dot from './dot';
 
 const variants = {
   keyValue,
   primary,
+  dot,
 };
 
 export default variants;
