+++ src/theme/badges/primary.ts
@@ -1,0 +1,29 @@
+const variants = {
+  bold: {
+    color: 'white',
+
+    bg: 'gray',
+  },
+  fail: {
+    color: 'white',
+
+    bg: 'labels.red',
+  },
+  label: {
+    color: 'white',
+
+    bg: 'black',
+  },
+  light: {
+    color: 'black',
+
+    bg: 'grayShade3',
+  },
+  success: {
+    color: 'white',
+
+    bg: 'labels.green',
+  },
+};
+
+export default variants;
