--- src/theme/badges/primary/bold.ts
@@ -1,5 +1,0 @@
-export default {
-  color: 'white',
-
-  bg: 'gray',
-};
