--- src/theme/badges/primary/fail.ts
@@ -1,5 +1,0 @@
-export default {
-  color: 'white',
-
-  bg: 'labels.red',
-};
