--- src/theme/badges/primary/index.ts
@@ -1,15 +1,0 @@
-import bold from './bold';
-import fail from './fail';
-import label from './label';
-import light from './light';
-import success from './success';
-
-const variants = {
-  bold,
-  fail,
-  label,
-  light,
-  success,
-};
-
-export default variants;
