--- src/theme/badges/primary/label.ts
@@ -1,5 +1,0 @@
-export default {
-  color: 'white',
-
-  bg: 'black',
-};
