--- src/theme/badges/primary/light.ts
@@ -1,5 +1,0 @@
-export default {
-  color: 'black',
-
-  bg: 'grayShade3',
-};
