--- src/theme/badges/primary/success.ts
@@ -1,5 +1,0 @@
-export default {
-  color: 'white',
-
-  bg: 'labels.green',
-};
